pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "STONKZILLA"
version = "0.1.1"
description = "Market price and indicator plotter."
readme = "README.md"
requires-python = ">=3.13.3"
dependencies = [
"annotated-types==0.7.0; python_version >= '3.8'",
"astroid==3.3.9; python_full_version >= '3.9.0'",
"beautifulsoup4==4.13.4; python_full_version >= '3.7.0'",
"certifi==2025.4.26; python_version >= '3.6'",
"cffi==1.17.1; python_version >= '3.8'",
"charset-normalizer==3.4.2; python_version >= '3.7'",
"click==8.1.8; python_version >= '3.7'",
"contourpy==1.3.2; python_version >= '3.10'",
"curl-cffi==0.10.0; python_version >= '3.9'",
"cycler==0.12.1; python_version >= '3.8'",
"dill==0.4.0; python_version >= '3.8'",
"fonttools==4.57.0; python_version >= '3.8'",
"frozendict==2.4.6; python_version >= '3.6'",
"idna==3.10; python_version >= '3.6'",
"isort==6.0.1; python_full_version >= '3.9.0'",
"kiwisolver==1.4.8; python_version >= '3.10'",
"matplotlib==3.10.1; python_version >= '3.10'",
"mccabe==0.7.0; python_version >= '3.6'",
"multitasking==0.0.11",
"numpy==2.2.5; python_version >= '3.10'",
"packaging==25.0; python_version >= '3.8'",
"pandas==2.2.3; python_version >= '3.9'",
"peewee==3.18.1",
"pillow==11.2.1; python_version >= '3.9'",
"platformdirs==4.3.8; python_version >= '3.9'",
"pycparser==2.22; python_version >= '3.8'",
"pydantic==2.11.4; python_version >= '3.9'",
"pydantic-core==2.33.2; python_version >= '3.9'",
"pyparsing==3.2.3; python_version >= '3.9'",
"python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'",
"pytz==2025.2",
"pyyaml==6.0.2; python_version >= '3.8'",
"requests==2.32.3; python_version >= '3.8'",
"six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'",
"soupsieve==2.7; python_version >= '3.8'",
"tomlkit==0.13.2; python_version >= '3.8'",
"typing-extensions==4.13.2; python_version >= '3.8'",
"typing-inspection==0.4.0; python_version >= '3.9'",
"tzdata==2025.2; python_version >= '2'",
"urllib3==2.4.0; python_version >= '3.9'",
"yfinance==0.2.58"

]

license = { text = "MIT" }

authors = [
  { name = "Jakub Gąsior", email = "jakubgasior72@gmail.com" }
]

[project.scripts]
stonkzilla = "stonkzilla.main:main"

[tool.setuptools]
packages = [
  "stonkzilla",
  "stonkzilla.cli",
  "stonkzilla.data_sources",
  "stonkzilla.plots",
  "stonkzilla.indicators"
  ]


[tool.setuptools.package-dir]
"" = "."